Rename 'components/MonitoringItemsOverviewTableMedicalSystem.vue' to 'components/index/CardsMonitoring/MonitoringItemsOverview/MedicalSystem.vue'
