+++ components/index/CardsMonitoring/MonitoringItemsOverview/MedicalSystem.vue
